Database Behavior Fixes
Fixes to Main Activity Layout and ListView Adapter Layout to show
responses better and allow for scrolling of long list view entries.
Fixed for repeat delay in query of the database and use of the timing
values recorded there to determine if it’s been long enough to need to
send another message.