Fix crashers. Externalize Strings
Fixed the crashers I was seeing earlier. Needed to make sure we do
things in the right order and initialize stuff before we use it. Also
added gating conditions so we don’t try to use Google Services if
they’re not available.
Started moving more strings out of prefs panels and into the strings
file.